main/gcc6: fix missing includes with newer isl
Fix the following error:
gcc/graphite-isl-ast-to-gimple.c:114:7: error: 'isl_id_free' was not declared in this scope; did you mean 'isl_aff_free'?

Found the patch here:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=86724

Note that Alpine edge has isl 0.18 and not 0.20, but the error message
is the same and the patch works.
[ci:skip-vercheck]: no need to rebuild existing packages for this

Related: https://builds.sr.ht/~postmarketos/job/489478