huawei-g7: new device (Huawei Ascend G7) (MR 2287)
The Huawei Ascend G7 is based on Qualcomm MSM8916 and therefore makes
use of linux-postmarketos-qcom-msm8916 as a close-to-mainline kernel.

Initially supported features:
  - USB
  - Storage (eMMC, SD card)
  - Display
  - Touchscreen, Buttons, Vibrator
  - Notification LED
  - WiFi/Bluetooth
  - Accelerometer, Magnetometer, Proximity
  - Audio
  - NFC

The modem needs more work to work around strange Huawei firmware mess.

Note that the Huawei shipped an outdated firmware version that does
not support booting aarch64 kernels (like used in this device port).
However, the device does not verify firmware signatures so it is
possible to flash firmware from other devices. This is mostly
described in the wiki, but unfortunately it's quite complex.

However, the aarch64 version of linux-postmarketos-qcom-msm8916
receives much more testing so I would like to avoid adding more
armv7 MSM8916 devices unless absolutely necessary.

For pointercal:
  Screen dimension 720x1280
  Touchscreen dimension (reported by evtest) 1100x1900
    720/1100*65536 = ~42896
    1280/1900*65536 = ~44151
  Offset doesn't seem to be needed.

[ci:skip-build] already built successfully in CI