linux-purism-librem5: upgrade to 5.11.15 (MR 2140)
Upstream changelog:
https://source.puri.sm/Librem5/linux-next/-/commit/4e76143d2ec5694ec09da20c9e702ed2f0dbdb3f

Note, previous wifi driver behavior is not changed. The redpine module
still expects the fw to be in /lib/firmware on boot, and won't use the
fw on the card.. Theoretically this kernel version now supports both
loading fw from rootfs into ram AND using fw burned on the chip, but
since I haven't tested that well yet I'm leaving this as-is.

[ci:skip-build]: already built successfully in CI