soc-qcom-sdm845: 5.14 feedbackd fix and modemmanager suspend/resume support (MR 2416)
Fix the feedbackd rule as the haptics modules has been renamed.

Although SDM845 doesn't really suspend well, MM needs suspend
and resume support when the device attempts to suspend. Without
which MM will crash and wont resume properly when a suspend
action is triggered.

Depend on msm-modem-uim-selection to configure to modem for modemmanager