treewide: firmware: use pmb:cross-native/!tracdeps (MR 2468)
Add pmb:cross-native/!tracedeps to applicable firmware packages and
secure both with a test.

This sets the pmb:cross-native and !tracedeps options on all firmware-*
packages that are compatible with the native compilation method. A unit
test ensures the presence of both options while maintaining a list of
exempted packages.

Fixes: #718
[ci:ignore-count] [ci:skip-vercheck] [ci:skip-build]