main/postmarketos-mkinitfs: replace udhcpd with unudhcpd (MR 2698)
unudhcpd is a lightweight dhcp server that effectively implements static
assignment of a single IP on an interface regardless of client MAC, with
no lease. The end result is that a system connected to a pmOS device via
usb networking should always get an IP, and always get one very quickly.

This adds a new dependency on this dhcp server, and starts it in initfs.

Fixes #1199

[ci:skip-build] already built successfully in CI