main/postmarketos-ui-*: link to UI pkg reference (MR 2683)
Let every postmarketos-ui-* package have a link on top that points to
the new UI package reference. Do not put the link in <>, as it was done
with other references, since there's no benefit to it. It doesn't look
better and it doesn't make it easier to click the link or something, I
just did this initially because I saw it in GPL license headers.

[ci:skip-vercheck] [ci:ignore-count]