tolino-shine2hd: new device (MR 2797)
Basic support for the Tolino Shine 2 HD ebook reader.
Waveform handling shares the same oddities as for the Kobo Clara HD
but the file epdc_E060SCM.fw from the IMX6SL BSP from NXP does also work.
The existing waveform from SD card is used.

Preferred way to install is to clone the internal SD, keeping the original
as a backup, install on that clone.

U-Boot is a more recent one than the factory u-boot, so devicetrees are nicely
supported. Kernel is the near mainline kernel also used by the Kobo Clara HD.

About Wifi: Handling of the configuration file for boards with brcmfmac is not
supported yet.

[ci:skip-build] already built successfully in CI