firmware-samsung-m0: fix wifi on mainline, and add bt fw (MR 2873)
With this wifi seem to work reliably, and it is possibly to for
example connect to a bluetooth speaker and play music.

For wifi, issue seem to be with the nvram_net.txt file, changing only
that one to a newer one makes wifi work.  Here, both the .txt and .bin
file are updated, as the old files are from 2012, while the new
ones have a datestamp from 2019.