main/postmarketos-ui-plasma-mobile: update deps (MR 2938)
Add and rearrange dependencies with new apps, and fix angelfish package
name.

Adds the following packages to extras:
- alligator (rss reader)
- kasts (podcasts)
- tokodon (mastodon client)

Move neochat from a required to an extra dependency, since it's a
matrix client (which many people may not need).

Add plasma-mobile-sounds to main dependency, so that kclock (and dialer
in the future) have sounds to choose from.

Rename plasma-angelfish to angelfish dependency since plasma-angelfish
was renamed in alpine at some point.

Remove the part of the description of the package that says numeric
passwords are required (no longer the case).