temp/dino: rebuild against new libicuuc.so (MR 3093)
Fix for:
  missing depend(s): so:libicuuc.so.70

While at it, I've checked if this is the most recent version of this
mobile fork, and what the upstream stance on this is. Meanwhile there's
a feature branch in the official dino.git repository, so I've adjusted
the APKBUILD. The commit is still the same, and apparently upstream
won't merge it, but rather switch the whole application to libadwaita.

Related: https://github.com/dino/dino/issues/178
Related: https://github.com/dino/dino/issues/1182

[ci:skip-build] already built successfully in CI