temp/epiphany: drop (MR 3098)
Helps #1479

The Disable-fullscreen patch from Purism is in principle useful
in mobile, as exiting fullscreen is at least troublesome. However,
the patch has not been upstreamed and, although we could carry it
in alpine for the 42 cycle, the 43 cycle will come with GTK4 support
and the patch will not rebase easily. Therefore, instead of waiting
6 months to drop, lets just drop it now.

[ci:skip-build] already built successfully in CI