main/postmarketos-mkinitfs: provide only a virtual package (MR 3228)
Fixes 95a390435e67f190ce28cee27c9c3bbaab379a8e

The previous `provides="mkinifs=3.6.0"` was a requirement because alpine's
kernels had a minimum dependency in `mkinitfs>=3.6.0` to make sure that
support for compressed modules existed. That requirement is no longer
in place[1], since the dependency is now on the `initramfs-generator`
virtual package. Also, since `initramfs-generator` is a virtual package
and not a real package, versioning it might mess up APK. The documentation
explictly states that virtual packages are those with unversioned provides
and a provider_priority. So to avoid having problems in the future, let's
strictly stick to it.

[1] https://gitlab.alpinelinux.org/alpine/aports/-/commit/4fac61c826cac81af72e095b645098900d3094a1