device-kobo-clara: remove mesa-dri-gallium dep (MR 3371)
Having mesa-dri-gallium installed causes Xorg to crash. See #1658.
Furthermore, display works without at least with near-mainline kernel,
tested with osk-sdl and xfce4.
So if specific applications depend on it, it should be added as a
dependency there.

This reverts MR 3306

This is just a hotfix, Xorg/mesa should be tamed to not crash here
but analyzing that costs time.