main/postmarketos-ui-{mate,xfce4}: replace pulseaudio-utils by pulseaudio (MR 3479)
Actually package pulseaudio is needed instead of pulseaudio-utils.
Package pulseaudio-utils then gets drawn in by postmarketos-base-ui
sub-package pulseaudio.

For Xfce4 this is a cosmetic change because pulseaudio gets installed
as a dependency of pavucontol already. Keeping pulseaudio in the
_pmb_reccomends list makes sense to be clear which pieces are needed.

For MATE an explicit installation of pulseaudio is needed because it
doesn't use pavucontrol.

[ci:skip-vercheck]
[ci:skip-build]: already built successfully in CI