main/postmarketos-ui-plasma-mobile: remove firefox (MR 3499)
We already ship Angelfish and shipping 2 browser takes up a lot of
space. Besides Angelfish feels more native in Plasma Mobile (being baed
on Qt and made for mobile) and has a built-in adblocker nowadays so it
seems like the better option.

urfkill hasn't been used for quite a while, no need to keep it around

Motivation for this change is that the image builds of the UI are
currently running out of space so we need to decrease the image size