main/postmarketos-hidden-desktop-entries: remove (MR 4086)
GNOME Console can now launch command-line applications from .desktop
files, so they are no longer "useless" in Phosh.

gnome-extensions-app hasn't been pulled in automatically in Phosh for
long. If users install this manually, it would be confusing for it to
not show up. The app itself has an error message that explains they
need to be logged into GNOME for it to work when launched in e.g.
Phosh, so I think that should take care of UX concerns. Additionally,
this package is installed in gnome-mobile as well, where
gnome-extensions-app actually is useful.

firefox-safe-mode has been removed in Alpine:
 - https://git.alpinelinux.org/aports/commit/community/firefox?id=438ee43621e1a83c8309bf7e5e14ba9fce0532ff
 - https://git.alpinelinux.org/aports/commit/community/firefox-esr?id=0f324c8b39ab08cc3f504e4ae959425bec472791

Considering all of this, postmarketos-hidden-desktop-entries no longer
serves a purpose, so let's remove it.