linux-google-veryron: enable binfmt_misc as kernel module (MR 4115)
This allows the user (upon loading the module) to handle non-native
binaries transparently - after registering the magic number and the
command for handling execution, the programs can be ran as if compiled
for the host architecture. (Similarly how the shebang line allows
the interpreter to be specified for scirpts).

Common use-cases include

* Using qemu user-space emulation to run
  x86 binaries on ARM (or the other way around), or simply running
  64-bit binaries on 32-bit architectures.
* Allow Java programs to be run transparently.
* Creating cross-architecture chroots (for cross-compilation e.g.)