main/sxmo-common: Add a ghost package to fix upgrades (MR 4259)
This old package had a pre-deinstall script to remove the necessary Sxmo
daemons. We move those daemon management to the -dwm and -sway dedicated
postmarketos packages.

The problem is that apk can eventually install/upgrade the -dwm or -sway
package before removing sxmo-common. This cause a deactivation of the
daemons.

To prevent this, we add back an empty sxmo-common to take place of the
old one. This one does not contain the deinstall scripts. We will
remove this in two stable release.