main/devicepkg-dev: use modules-initfs to install modules into initramfs (MR 4169)
Since the modules are going to be installed into the initfs, it makes
sense to use a more verbose naming for the file. Otherwise there's the
risk of getting confused with other type of module files that exist in
device packages.

Fixes 8725dbbcee15a506ec980c3b42306964ab04c19e