huawei-kiwi: add GR5 (2016) (MR 4302)
GR5 is for international markets, and it's also kiwi.
The existing pmaports, lk2nd and kernel works can be used directly
without problems.

There is nothing different found yet. The 2 phones seem to be identical
except that GR5 is not branded as Honor series and it has no Honor logo
on the chassis.